backend/requirements.txt
# Core API
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
requests==2.31.0
httpx==0.27.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.30

# Data & ML (La partie critique)
# Pandas 2.x est nécessaire pour Python 3.12
pandas==2.2.2
numpy==1.26.4
# Scikit-learn 1.4+ gère mieux les validateurs récents, mais on fige pour éviter les surprises
scikit-learn==1.4.2
# XGBoost 2.0+ a une meilleure interface sklearn
xgboost==2.0.3
joblib==1.4.2

# Testing
pytest==8.2.0
pytest-mock==3.14.0